+++ src/WrapperApp/components/Results/EstimatorTab/EstimatorTab.tsx
@@ -1,6 +1,6 @@
 import { Box, CircularProgress, useTheme } from '@mui/material';
 
-import { generateGraphs } from '../../../../JsRoot/NewGraphData';
+import { generateGraphs } from '../../../../JsRoot/GraphData';
 import { FullSimulationData } from '../../../../services/ShSimulatorService';
 import { EstimatorResults } from '../ResultsPanel';
 import TablePage0D from '../ResultsTable';
